[polish] hello world C
