+++ src/examples/hello-world/c/helloWorld.md
@@ -1,1 +1,3 @@
-clang --target=wasm32 --no-standard-libraries -Wl,--export-all -Wl,--no-entry -o helloWorld.wasm helloWorld.c
+# Hello World (C)
+
+Compile with `clang --target=wasm32 --no-standard-libraries -Wl,--export-all -Wl,--no-entry -o helloWorld.wasm helloWorld.c`
