+++ tsconfig.json
@@ -25,6 +25,6 @@
     "src/**/*.ts",
     "src/**/*.tsx",
     ".next/types/**/*.ts"
-,],
+  ],
   "exclude": ["node_modules"]
 }
